requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --universal --output-file requirements.txt
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via langchain
aiosignal==1.3.2
    # via aiohttp
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via httpx
attrs==24.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
bcrypt==4.2.1
    # via -r requirements.in
blinker==1.9.0
    # via streamlit
cachetools==5.5.0
    # via streamlit
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via streamlit
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
cryptography==44.0.0
    # via pyjwt
deprecated==1.2.15
    # via pygithub
et-xmlfile==2.0.0
    # via openpyxl
faiss-cpu==1.9.0.post1
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
greenlet==3.1.1 ; (python_full_version < '3.13' and platform_machine == 'AMD64') or (python_full_version < '3.13' and platform_machine == 'WIN32') or (python_full_version < '3.13' and platform_machine == 'aarch64') or (python_full_version < '3.13' and platform_machine == 'amd64') or (python_full_version < '3.13' and platform_machine == 'ppc64le') or (python_full_version < '3.13' and platform_machine == 'win32') or (python_full_version < '3.13' and platform_machine == 'x86_64')
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via langsmith
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.5
    # via
    #   altair
    #   pydeck
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
langchain==0.3.14
    # via -r requirements.in
langchain-core==0.3.29
    # via
    #   langchain
    #   langchain-text-splitters
langchain-text-splitters==0.3.5
    # via langchain
langsmith==0.2.10
    # via
    #   langchain
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
narwhals==1.21.1
    # via altair
numpy==2.2.1
    # via
    #   faiss-cpu
    #   langchain
    #   pandas
    #   pydeck
    #   streamlit
openpyxl==3.1.5
    # via -r requirements.in
orjson==3.10.14 ; platform_python_implementation != 'PyPy'
    # via langsmith
packaging==24.2
    # via
    #   altair
    #   faiss-cpu
    #   langchain-core
    #   streamlit
pandas==2.2.3
    # via
    #   -r requirements.in
    #   streamlit
pillow==11.1.0
    # via
    #   -r requirements.in
    #   streamlit
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.3
    # via streamlit
pyarrow==18.1.0
    # via streamlit
pycparser==2.22
    # via cffi
pydantic==2.10.4
    # via
    #   langchain
    #   langchain-core
    #   langsmith
pydantic-core==2.27.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygithub==2.5.0
    # via -r requirements.in
pygments==2.19.1
    # via rich
pyjwt==2.10.1
    # via pygithub
pynacl==1.5.0
    # via pygithub
pypdf==5.1.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   langchain
    #   langsmith
    #   pygithub
    #   requests-toolbelt
    #   streamlit
requests-toolbelt==1.0.0
    # via langsmith
rich==13.9.4
    # via streamlit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via langchain
streamlit==1.41.1
    # via -r requirements.in
tenacity==9.0.0
    # via
    #   langchain
    #   langchain-core
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
typing-extensions==4.12.2
    # via
    #   altair
    #   anyio
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   pygithub
    #   sqlalchemy
    #   streamlit
tzdata==2024.2
    # via pandas
urllib3==2.3.0
    # via
    #   pygithub
    #   requests
watchdog==6.0.0 ; sys_platform != 'darwin'
    # via streamlit
wrapt==1.17.0
    # via deprecated
xlsxwriter==3.2.0
    # via -r requirements.in
yarl==1.18.3
    # via aiohttp
